refactor: 디렉토리명 변경 및 파일구조 이동
